requirements.txt
gradio
torch
torchvision
deoldify
Pillow
